requirements.txt
arcade==2.4.3
attrs==20.3.0
cffi==1.14.4
llvmlite==0.35.0
numba==0.52.0
numpy==1.19.3
Pillow==8.0.1
pycparser==2.20
pyglet==1.5.13
pymunk==6.0.0
pytiled-parser==0.9.4a3
PyYAML==5.3.1
